+++ NEWS.md
@@ -1,3 +1,7 @@
+DALEX 0.4.2
+----------------------------------------------------------------
+* New `yhat` functions for `mlr`, `h2o` and `caret` packages (added by Szymon).
+
 DALEX 0.4.1
 ----------------------------------------------------------------
 * `plot.variable_importance_explainer()` has now  `desc_sorting` argument. If FALSE then variable importance will be sorted in an increasing order ([#41](https://github.com/pbiecek/DALEX/issues/41)).